move Contact model to models folder
