+++ models/index.js
@@ -1,0 +1,1 @@
+export { default as Contact } from "./contactModelSchema.js";
