--- schemas/contactsModelSchema.js
@@ -1,27 +1,0 @@
-import { Schema, model } from "mongoose";
-
-const contactSchema = new Schema({
-  name: {
-    type: String,
-    required: [true, "Set name for contact"],
-  },
-  email: {
-    type: String,
-  },
-  phone: {
-    type: String,
-  },
-  favorite: {
-    type: Boolean,
-    default: false,
-  },
-});
-
-const Contact = model("contact", contactSchema);
-
-contactSchema.post("save", (er, data, next) => {
-  er.status = 400;
-  next();
-});
-
-export default Contact;
