+++ schemas/index.js
@@ -1,2 +1,1 @@
 export { default as contactJoiSchemas } from "./contactJoiSchemas.js";
-export {default as Contact} from './contactsModelSchema.js'
